+++ src/common/docBuilder/docBuilderPackageXml.ts
@@ -6,6 +6,7 @@ import { uxLog } from '../utils/index.js';
 import { countPackageXmlItems, parsePackageXmlFile } from '../utils/xmlUtils.js';
 import { SalesforceSetupUrlBuilder } from './docUtils.js';
 import { CONSTANTS } from '../../config/index.js';
+import { prettifyFieldName } from '../utils/flowVisualiser/nodeFormatUtils.js';
 
 export class DocBuilderPackageXML {
 
@@ -65,6 +66,8 @@ export class DocBuilderPackageXML {
         '',
         packageXmlDefinition.description,
         '',
+        '<div id="jstree-container"></div>',
+        '',
         `Metadatas: ${nbItems}`,
         ''
       ]);
@@ -73,6 +76,8 @@ export class DocBuilderPackageXML {
       // Header
       mdLines.push(...[
         `## Content of ${path.basename(inputFile)}`,
+        '',
+        '<div id="jstree-container"></div>',
         '',
         `Metadatas: ${nbItems}`,
         ''
@@ -106,8 +111,16 @@ export class DocBuilderPackageXML {
 
     // Write output file
     await fs.writeFile(outputFile, mdLines.join("\n") + "\n");
-
     uxLog(this, c.green(`Successfully generated ${path.basename(inputFile)} documentation into ${outputFile}`));
+
+    const jsonTree = await this.generateJsonTree(metadataTypes, packageXmlContent);
+    if (jsonTree) {
+      const packageXmlFileName = path.basename(outputFile, ".md");
+      const jsonFile = `./docs/json/root-${packageXmlFileName}.json`;
+      await fs.ensureDir(path.dirname(jsonFile));
+      await fs.writeFile(jsonFile, JSON.stringify(jsonTree, null, 2));
+      uxLog(this, c.green(`Successfully generated ${packageXmlFileName} JSON into ${jsonFile}`));
+    }
 
     return outputFile;
   }
@@ -147,4 +160,70 @@ export class DocBuilderPackageXML {
     ];
   }
 
+  // Generate json for display with jsTree npm library 
+  public static async generateJsonTree(metadataTypes: any, packageXmlContent: any): Promise<any> {
+    const treeElements: any[] = [];
+    for (const metadataType of metadataTypes) {
+      const members = packageXmlContent[metadataType] || [];
+      members.sort();
+      const memberLengthLabel = members.length === 1 && members[0] === "*" ? "all" : members.length;
+      const typeRoot: any = {
+        text: prettifyFieldName(metadataType) + " (" + memberLengthLabel + ")",
+        icon: memberLengthLabel !== "all" ? "fa-solid fa-folder icon-blue" : "fa-solid fa-folder icon-warning",
+        a_attr: { href: null },
+        children: [],
+      }
+      if (memberLengthLabel !== "all") {
+        if (metadataType === "CustomField") {
+          // Sort custom fields by object name
+          DocBuilderPackageXML.createCustomFieldsTree(members, typeRoot);
+        }
+        else {
+          DocBuilderPackageXML.createMembersTree(members, typeRoot);
+        }
+      }
+      treeElements.push(typeRoot);
+    }
+    return treeElements;
+  }
+
+  private static createCustomFieldsTree(members: any, typeRoot: any) {
+    const elementsByObject: any = [];
+    for (const element of members) {
+      const objectName = element.split('.')[0];
+      if (!elementsByObject[objectName]) {
+        elementsByObject[objectName] = [];
+      }
+      elementsByObject[objectName].push(element);
+    }
+    // Create object nodes and fields as children
+    for (const objectName of Object.keys(elementsByObject)) {
+      const objectNode: any = {
+        text: objectName + " (" + elementsByObject[objectName].length + ")",
+        icon: "fa-solid fa-folder icon-blue",
+        a_attr: { href: null },
+        children: [],
+      };
+      for (const element of elementsByObject[objectName]) {
+        const subElement: any = {
+          text: element,
+          icon: "fa-solid fa-circle-check icon-success",
+          a_attr: { href: null },
+        };
+        objectNode.children.push(subElement);
+      }
+      typeRoot.children.push(objectNode);
+    }
+  }
+
+  private static createMembersTree(members: any, typeRoot: any) {
+    for (const member of members) {
+      const subElement: any = {
+        text: member,
+        icon: "fa-solid fa-circle-check icon-success",
+        a_attr: { href: null },
+      };
+      typeRoot.children.push(subElement);
+    }
+  }
 }